refactor(UserService): 클래스 인터페이스 추출
