+++ app/src/main/java/camp/nextstep/service/UserService.java
@@ -1,46 +1,13 @@
 package camp.nextstep.service;
 
-import camp.nextstep.dao.UserDao;
-import camp.nextstep.dao.UserHistoryDao;
 import camp.nextstep.domain.User;
-import camp.nextstep.domain.UserHistory;
-import camp.nextstep.transaction.support.TransactionTemplate;
-import com.interface21.beans.factory.annotation.Autowired;
-import com.interface21.context.stereotype.Service;
 
-@Service
-public class UserService {
+public interface UserService {
+    User findByAccount(String account);
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final TransactionTemplate transactionTemplate;
+    User findById(long id);
 
-    @Autowired
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao, final TransactionTemplate transactionTemplate) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.transactionTemplate = transactionTemplate;
-    }
+    void save(User user);
 
-    public User findByAccount(final String account) {
-        return userDao.findByAccount(account);
-    }
-
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void save(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        transactionTemplate.run(() -> {
-            final var user = findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-        });
-    }
-
+    void changePassword(long id, String newPassword, String createBy);
 }
